Auth added

+++ backend/controllers/goalController.js
@@ -1,17 +1,16 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const Users = require('../models/userModel')
 
 //  @desc   Get goals
 //  @route  GET /api/goals  
 //  @access Private
 const getGoals = asyncHandler(async (req, res) => {
 
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
 
     res.status(200).json(goals)
-
-    res.status(200).json({message: 'Get goals'})
 })
 
 //  @desc   Set goal
@@ -24,7 +23,8 @@ const setGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create( {
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     
     console.log(req.body)
@@ -44,6 +44,20 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not gound')
     }
 
+    const user = await Users.findById(req.user.id)
+
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
 const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
 })
@@ -61,7 +75,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
 
     if(!goal) {
         res.status(400)
-        throw new Error('Goal not gound')
+        throw new Error('Goal not found')
+    }
+
+    const user = await Users.findById(req.user.id)
+
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await goal.deleteOne()
